requirements.txt
python-dotenv
slack-bolt
slackclient
slackeventsapi
flask
